sampling_setup: split the tests out — 518 LoC over the 500 cap
The min_p/top_n_sigma precedence work pushed
`spark-server/src/api/chat/sampling_setup.rs` to 518 lines, and
`.github/workflows/file-size-cap.yml` does not excuse it: the allow-list
is 111 entries and this file is not one of them. Adding it there would be
the wrong fix — the allow-list exists for files that were already over
when the cap landed, not for new growth.

The 111-line `#[cfg(test)] mod tests` block moves to a sibling
`sampling_setup_tests.rs`, declared the way `msg_entry_tests.rs` already
is in this same directory:

    #[cfg(test)]
    #[path = "sampling_setup_tests.rs"]
    mod sampling_setup_tests;

Parent 518 -> 410, tests file 117. No behaviour change; the resolution
logic is untouched.

One trap worth naming, because it has now cost time twice on this branch:
moving a `mod tests` body into its own file changes what `super` means.
Inside `sampling_setup.rs` the block was one module deep, so `super::foo`
reached the parent's items. As a `#[path]` sibling it is still a child
module — but its `use` lines were resolving names that live in
`api::chat`, not in `sampling_setup`, so they need `super::super::`.
`cargo clippy --all-targets` catches it (E0432); a plain `cargo build`
does not, because it never compiles the test cfg.

Verified locally against the workflow's own allow-list (a script that
parses `allow_list=( ... )` out of the yml rather than guessing):
PASS, every .rs under crates/ is <=500 LoC or allow-listed.
`cargo fmt --all --check` clean, `cargo clippy -p spark-server
--all-targets` clean, `cargo test -p spark-server --lib` 104 passed.

Note on the one red test in this crate: `cargo test -p spark-server`
also reports `the_baked_attestation_reproduces_from_the_tree` failing on
`gb10/qwen3.8-27b/nvfp4`. That is not from this change and not from the
longcat kernel symlinks. A target's closure hash covers
`kernels/<hw>/common/*.cu` plus its own model dir; commit 0c8222fa added
`kernels/gb10/common/moe_topk_softmax_bias.cu` for the zero-expert
router, which is in *every* gb10 target's input set, so every hash baked
before it is stale. The build re-bakes on a real kernel build. The
symlink commit fe2c76a5 touched only
`kernels/gb10/longcat-flash-lite/nvfp4/`, which is not in qwen3.8-27b's
set at all.
